Move nodeID and time range filtering from LineChart to FeatureView
